requirements.txt
Django==2.1.7
django-ckeditor==5.6.1
django-jet==2.0.1
django-js-asset==1.2.1
django-modeltranslation==0.13.1
django-widget-tweaks==1.4.3
feedparser==5.2.1
Pillow==5.4.1
#pkg-resources==0.0.0
pytz==2018.9
sqlparse==0.3.0
